refactor (api): separate local and production entrypoints
