--- .env
@@ -1,10 +1,0 @@
-# Everytime we change this file bash shell must be restarted. bash automatically
-# loads the .env file only once when it starts.
-
-# Frontend web app configuration
-WEB_HOST="localhost"
-WEB_PORT=5173
-
-# Backend API server configuration where the api server shall listenn
-VITE_API_HOST="localhost"
-VITE_API_PORT=8181
