+++ README.md
@@ -114,13 +114,14 @@ You can also debug with VS Code using the following `.vscode/launch.json` config
 ```
 
 ### Build for production
-Don't forget to set environment variables for API endpoints before building. Take a look inside `.env`, and set the ones prefixed with `VITE_`.
 
 ```bash
 npm run build
 ```
 
 The production build can be found in the `dist` directory.
+
+`web/public/config.json` is used for fetching the `api url`, and is set to localhost by default, but this file is exchanged during deployment to the productin endpoint.
 
 ### Run Unit Tests (React Components)
 
@@ -137,7 +138,7 @@ cd api
 
 ### Install FastAPI
 
-`FastAPI` is the python API framework for creating APIs, and `Uvicorn` is the server program that hosts the FastAPI application.
+`FastAPI` is the python API framework for creating APIs, and `Uvicorn` is the server program that hosts the FastAPI application when running locally.
 
 ```bash
 python3 -m venv .venv
@@ -145,12 +146,13 @@ source .venv/bin/activate
 pip install -r requirments.txt
 ```
 
-### Run FastAPI server
-Don't forget to override the default environment variables in `.env` if you are running in production. This is done by setting environment variables with the same name. Currently, this is used for configuring the server endpoint, and also to allow CORS (Cross-Origin Resource sharing) for the web app.
+### Run FastAPI server locally
 
 ```bash
 python main.py
 ```
+
+In production, `lambda.py` defines the handler for the lambda function. 
 
 ---
