+++ api/config.py
@@ -1,13 +1,12 @@
-import os
-from dotenv import load_dotenv
+import json
+from pathlib import Path
 
-load_dotenv("../.env")
+PROJECT_DIR_PATH = Path(__file__).resolve().parents[1]
 
-# backend api
-API_HOST = os.getenv("VITE_API_HOST")
-API_PORT = int(os.getenv("VITE_API_PORT"))
+# backend api server
+CONFIG_FILE_PATH = PROJECT_DIR_PATH / "web" / "public" / "config.json"
+CONFIG = json.loads(CONFIG_FILE_PATH.read_text())
 
-# fronetend web app
-WEB_HOST = os.getenv("WEB_HOST")
-WEB_PORT = int(os.getenv("WEB_PORT"))
-WEB_URL = f"http://{WEB_HOST}:{WEB_PORT}"
+API_URL = CONFIG["apiUrl"]
+API_HOST = API_URL.split(":")[1].replace("//", "")
+API_PORT = int(API_URL.split(":")[2])
