+++ api/data.py
@@ -1,0 +1,12 @@
+QUESTIONS = [
+    {
+        "text": "What kind of storage is AWS S3",
+        "answers": ["Object storage", "SQL database", "Block storage", "Document database"],
+        "correctAnswer": 0,
+    },
+    {
+        "text": "What does EC2 mean",
+        "answers": ["Elastic Container Creator", "Elastic Compute Cloud", "Ephemeral Code Catalog", "Error Cloud 2"],
+        "correctAnswer": 1,
+    },
+]
