+++ api/lambda.py
@@ -1,0 +1,6 @@
+"""Entrypoint for lambda function in aws."""
+import mangum
+from api import app
+
+
+handler = mangum(app)
