+++ api/question.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import List
+
+class Question(BaseModel):
+    text: str
+    answers: List[str]
+    correctAnswer: int
