+++ api/requirements.txt
@@ -7,7 +7,6 @@ idna==3.10
 mangum==0.19.0
 pydantic==2.9.2
 pydantic_core==2.23.4
-python-dotenv==1.0.1
 sniffio==1.3.1
 starlette==0.41.0
 typing_extensions==4.12.2
